(PC-12725) feat(Styling): ajout de styled-components pour le Web
