+++ src/App.web.tsx
@@ -24,7 +24,7 @@ import { eventMonitoring } from 'libs/monitoring'
 import { useStartBatchNotification } from 'libs/notifications'
 import { SafeAreaProvider } from 'libs/react-native-save-area-provider'
 import { ReactQueryClientProvider } from 'libs/react-query/ReactQueryClientProvider'
-import { ThemeProvider } from 'libs/styled/ThemeProvider'
+import { ThemeProvider } from 'libs/styled'
 import { theme } from 'theme'
 import { LoadingPage } from 'ui/components/LoadingPage'
 import { SnackBarProvider } from 'ui/components/snackBar/SnackBarContext'
