+++ src/libs/styled/index.tsx
@@ -1,0 +1,1 @@
+export { ThemeProvider } from './native/ThemeProvider'
