+++ src/libs/styled/native/types.ts
@@ -1,0 +1,11 @@
+import { DefaultTheme } from 'styled-components/native'
+
+export type Computed = {
+  isMobileViewport?: boolean
+  isTabletViewport?: boolean
+  isDesktopViewport?: boolean
+  showTabBar: boolean
+  appContentWidth: number
+}
+
+export type ComputedTheme = DefaultTheme & Computed
