+++ src/libs/styled/web/types.web.ts
@@ -1,0 +1,6 @@
+import { DefaultTheme as DefaultThemeWeb } from 'styled-components'
+import { DefaultTheme } from 'styled-components/native'
+
+import { Computed } from '../native/types'
+
+export type ComputedTheme = DefaultThemeWeb & DefaultTheme & Computed
