+++ src/theme/styled.d.web.ts
@@ -1,0 +1,7 @@
+import 'styled-components/native'
+import { AppThemeType } from 'theme/index'
+
+declare module 'styled-components' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  export interface DefaultTheme extends AppThemeType {}
+}
